Move memory region structs to `boot_info` module
