+++ src/binary/mod.rs
@@ -1,6 +1,7 @@
-use crate::binary::legacy_memory_region::{LegacyFrameAllocator, LegacyMemoryRegion};
-use crate::boot_info::{BootInfo, FrameBuffer, FrameBufferInfo, TlsTemplate};
-use crate::memory_region::MemoryRegion;
+use crate::{
+    binary::legacy_memory_region::{LegacyFrameAllocator, LegacyMemoryRegion},
+    boot_info::{BootInfo, FrameBuffer, FrameBufferInfo, MemoryRegion, TlsTemplate},
+};
 use core::{
     mem::{self, MaybeUninit},
     slice,
