+++ src/boot_info.rs
@@ -1,4 +1,3 @@
-use crate::memory_region::MemoryRegion;
 use core::{ops, slice};
 
 /// This structure represents the information that the bootloader passes to the kernel.
@@ -102,6 +101,52 @@ impl From<MemoryRegions> for &'static mut [MemoryRegion] {
     fn from(regions: MemoryRegions) -> &'static mut [MemoryRegion] {
         unsafe { slice::from_raw_parts_mut(regions.ptr, regions.len) }
     }
+}
+
+/// Represent a physical memory region.
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[repr(C)]
+pub struct MemoryRegion {
+    /// The physical start address of the region.
+    pub start: u64,
+    /// The physical end address (exclusive) of the region.
+    pub end: u64,
+    /// The memory type of the memory region.
+    ///
+    /// Only [`Usable`][MemoryRegionKind::Usable] regions can be freely used.
+    pub kind: MemoryRegionKind,
+}
+
+impl MemoryRegion {
+    /// Creates a new empty memory region (with length 0).
+    pub const fn empty() -> Self {
+        MemoryRegion {
+            start: 0,
+            end: 0,
+            kind: MemoryRegionKind::Bootloader,
+        }
+    }
+}
+
+/// Represents the different types of memory.
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[non_exhaustive]
+#[repr(C)]
+pub enum MemoryRegionKind {
+    /// Unused conventional memory, can be used by the kernel.
+    Usable,
+    /// Memory mappings created by the bootloader, including the kernel and boot info mappings.
+    ///
+    /// This memory should _not_ be used by the kernel.
+    Bootloader,
+    /// An unknown memory region reported by the UEFI firmware.
+    ///
+    /// This should only be used if the UEFI memory type is known as usable.
+    UnknownUefi(u32),
+    /// An unknown memory region reported by the BIOS firmware.
+    ///
+    /// This should only be used if the BIOS memory type is known as usable.
+    UnknownBios(u32),
 }
 
 /// A pixel-based framebuffer that controls the screen output.
